requirements.txt
pinecone-client
openai
flask
psycopg2-binary

